refactor dump_index
